Add cancelling non-current jobs in a jobset
This allows users with the cancel jobs roles to cancel all (old) jobs
in a jobset. Previously this existed as an admin feature on a global
level. On the scale of hydra.nixos.org the global cleanup might be a bit
heavy in most situations. For jobsets such as haskell-updates,
staging-next, ... it might be good to have the ability to cancel all
outdated builds as those will not end up in a release channel anyway.